# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CPI Artifacts API
  version: 1.0.0
  description: API Telegram Mini App для сбора материалов мероприятий.
servers:
  - url: https://artifacts.62-113-105-225.sslip.io
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/uploads/init:
    post:
      tags: [uploads]
      summary: Создать simple или multipart S3-загрузку
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [submissionId, fileName, mimeType, sizeBytes]
              properties:
                submissionId: { type: string, format: uuid }
                fileName: { type: string, maxLength: 500 }
                mimeType: { type: string, maxLength: 200 }
                sizeBytes: { type: integer, minimum: 1 }
                lastModified: { type: integer, minimum: 0 }
      responses:
        '201':
          description: Параметры прямой загрузки
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadInit' }
        '413': { description: Файл превышает лимит или запрещён политикой }
  /api/v1/uploads/{artifactId}/part-url:
    get:
      tags: [uploads]
      summary: Короткоживущий URL отдельной multipart-части
      parameters:
        - { $ref: '#/components/parameters/ArtifactId' }
        - {
            name: partNumber,
            in: query,
            required: true,
            schema: { type: integer, minimum: 1, maximum: 10000 },
          }
      responses:
        '200':
          description: Presigned URL части
  /api/v1/uploads/{artifactId}/complete:
    post:
      tags: [uploads]
      summary: Идемпотентно завершить загрузку и поставить проверку в очередь
      parameters:
        - { $ref: '#/components/parameters/ArtifactId' }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                parts:
                  type: array
                  maxItems: 10000
                  items:
                    type: object
                    required: [partNumber, etag]
                    properties:
                      partNumber: { type: integer }
                      etag: { type: string }
      responses:
        '200':
          description: Метаданные артефакта
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Artifact' }
  /api/v1/uploads/{artifactId}/abort:
    post:
      tags: [uploads]
      summary: Отменить незавершённую загрузку
      parameters:
        - { $ref: '#/components/parameters/ArtifactId' }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '204': { description: Загрузка отменена }
  /api/v1/artifacts/{artifactId}/download:
    get:
      tags: [artifacts]
      summary: Защищённый временный URL готового файла
      parameters:
        - { $ref: '#/components/parameters/ArtifactId' }
      responses:
        '200':
          description: Presigned download URL
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cpi_artifacts_session
  parameters:
    CsrfToken:
      name: x-csrf-token
      in: header
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
...
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    UploadInit:
      type: object
      required: [artifactId, uploadType, expiresInSeconds]
      properties:
        artifactId: { type: string, format: uuid }
        uploadType: { enum: [simple, multipart] }
        uploadUrl: { type: string, format: uri }
        partSize: { type: integer }
        expiresInSeconds: { type: integer }
        alreadyCompleted: { type: boolean }
    Artifact:
      type: object
      properties:
        id: { type: string, format: uuid }
        submissionId: { type: string, format: uuid }
        eventId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        originalName: { type: string }
        displayName: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        checksumSha256: { type: [string, 'null'] }
        status:
          enum: [created, uploading, uploaded, verifying, ready, failed, quarantined, deleted]
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: {}
  responses:
...
    NotFound:
      description: Объект не найден или недоступен пользователю
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }